+++ src/lib/tvgPictureImpl.h
@@ -223,7 +223,7 @@ struct Picture::Impl
     }
 
 
-    float totalFrame()
+    uint32_t totalFrame()
     {
         return loader->totalFrame;
     }
